compress test files; add coverage of passthrough
